Fix static initialization, split more modules
