+++ src/modules/luaregister.h
@@ -5,14 +5,14 @@ class LuaScriptInterface;
 
 namespace tfs::lua {
 
-void registerModule(std::function<void(LuaScriptInterface&)> init);
+void registerModule(std::string_view moduleName, std::function<void(LuaScriptInterface&)> init);
 void importModules(LuaScriptInterface& lsi);
 
 } // namespace tfs::lua
 
-#define registerLuaModule(init) \
+#define registerLuaModule(moduleName, init) \
 	static auto __module__ = ([]() { \
-		tfs::lua::registerModule(init); \
+		tfs::lua::registerModule(moduleName, init); \
 		return nullptr; \
 	})();
