+++ src/modules/player.cpp
@@ -1,0 +1,2418 @@
+#include "../otpch.h"
+
+#include "../player.h"
+
+#include "../depotchest.h"
+#include "../game.h"
+#include "../inbox.h"
+#include "../iologindata.h"
+#include "../luaapi.h"
+#include "../luaenv.h"
+#include "../luameta.h"
+#include "../luascript.h"
+#include "../spectators.h"
+#include "../spells.h"
+#include "../storeinbox.h"
+#include "../vocation.h"
+#include "luaregister.h"
+
+extern Chat* g_chat;
+extern Game* g_game;
+extern Spells* g_spells;
+extern Vocations g_vocations;
+
+namespace {
+
+using namespace tfs::lua;
+
+int luaPlayerCreate(lua_State* L)
+{
+	// Player(id or guid or name or userdata)
+	Player* player;
+	if (isNumber(L, 2)) {
+		uint32_t id = getNumber<uint32_t>(L, 2);
+		if (id >= CREATURE_ID_MIN && id <= Player::playerIDLimit) {
+			player = g_game->getPlayerByID(id);
+		} else {
+			player = g_game->getPlayerByGUID(id);
+		}
+	} else if (lua_isstring(L, 2)) {
+		ReturnValue ret = g_game->getPlayerByNameWildcard(getString(L, 2), player);
+		if (ret != RETURNVALUE_NOERROR) {
+			lua_pushnil(L);
+			lua_pushnumber(L, ret);
+			return 2;
+		}
+	} else if (lua_isuserdata(L, 2)) {
+		if (getUserdataType(L, 2) != LuaData_Player) {
+			lua_pushnil(L);
+			return 1;
+		}
+		player = getUserdata<Player>(L, 2);
+	} else {
+		player = nullptr;
+	}
+
+	if (player) {
+		pushUserdata<Player>(L, player);
+		setMetatable(L, -1, "Player");
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerIsPlayer(lua_State* L)
+{
+	// player:isPlayer()
+	pushBoolean(L, getUserdata<const Player>(L, 1) != nullptr);
+	return 1;
+}
+
+int luaPlayerGetGuid(lua_State* L)
+{
+	// player:getGuid()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getGUID());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetIp(lua_State* L)
+{
+	// player:getIp()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getIP());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetAccountId(lua_State* L)
+{
+	// player:getAccountId()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getAccount());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetLastLoginSaved(lua_State* L)
+{
+	// player:getLastLoginSaved()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getLastLoginSaved());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetLastLogout(lua_State* L)
+{
+	// player:getLastLogout()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getLastLogout());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetAccountType(lua_State* L)
+{
+	// player:getAccountType()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getAccountType());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetAccountType(lua_State* L)
+{
+	// player:setAccountType(accountType)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->setAccountType(getNumber<AccountType_t>(L, 2));
+		IOLoginData::setAccountType(player->getAccount(), player->getAccountType());
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetCapacity(lua_State* L)
+{
+	// player:getCapacity()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getCapacity());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetCapacity(lua_State* L)
+{
+	// player:setCapacity(capacity)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->setCapacity(getNumber<uint32_t>(L, 2));
+		player->sendStats();
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetFreeCapacity(lua_State* L)
+{
+	// player:getFreeCapacity()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getFreeCapacity());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetDepotChest(lua_State* L)
+{
+	// player:getDepotChest(depotId[, autoCreate = false])
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint32_t depotId = getNumber<uint32_t>(L, 2);
+	bool autoCreate = getBoolean(L, 3, false);
+	DepotChest* depotChest = player->getDepotChest(depotId, autoCreate);
+	if (depotChest) {
+		pushUserdata<Item>(L, depotChest);
+		setItemMetatable(L, -1, depotChest);
+	} else {
+		pushBoolean(L, false);
+	}
+	return 1;
+}
+
+int luaPlayerGetInbox(lua_State* L)
+{
+	// player:getInbox()
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	Inbox* inbox = player->getInbox();
+	if (inbox) {
+		pushUserdata<Item>(L, inbox);
+		setItemMetatable(L, -1, inbox);
+	} else {
+		pushBoolean(L, false);
+	}
+	return 1;
+}
+
+int luaPlayerGetSkullTime(lua_State* L)
+{
+	// player:getSkullTime()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getSkullTicks());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetSkullTime(lua_State* L)
+{
+	// player:setSkullTime(skullTime)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->setSkullTicks(getNumber<int64_t>(L, 2));
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetDeathPenalty(lua_State* L)
+{
+	// player:getDeathPenalty()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getLostPercent() * 100);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetExperience(lua_State* L)
+{
+	// player:getExperience()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getExperience());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddExperience(lua_State* L)
+{
+	// player:addExperience(experience[, sendText = false])
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		uint64_t experience = getNumber<uint64_t>(L, 2);
+		bool sendText = getBoolean(L, 3, false);
+		player->addExperience(nullptr, experience, sendText);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerRemoveExperience(lua_State* L)
+{
+	// player:removeExperience(experience[, sendText = false])
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		uint64_t experience = getNumber<uint64_t>(L, 2);
+		bool sendText = getBoolean(L, 3, false);
+		player->removeExperience(experience, sendText);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetLevel(lua_State* L)
+{
+	// player:getLevel()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getLevel());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetMagicLevel(lua_State* L)
+{
+	// player:getMagicLevel()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getMagicLevel());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetBaseMagicLevel(lua_State* L)
+{
+	// player:getBaseMagicLevel()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getBaseMagicLevel());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetMana(lua_State* L)
+{
+	// player:getMana()
+	const Player* player = getUserdata<const Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getMana());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddMana(lua_State* L)
+{
+	// player:addMana(manaChange[, animationOnLoss = false])
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	int32_t manaChange = getNumber<int32_t>(L, 2);
+	bool animationOnLoss = getBoolean(L, 3, false);
+	if (!animationOnLoss && manaChange < 0) {
+		player->changeMana(manaChange);
+	} else {
+		CombatDamage damage;
+		damage.primary.value = manaChange;
+		damage.origin = ORIGIN_NONE;
+		g_game->combatChangeMana(nullptr, player, damage);
+	}
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerGetMaxMana(lua_State* L)
+{
+	// player:getMaxMana()
+	const Player* player = getUserdata<const Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getMaxMana());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetMaxMana(lua_State* L)
+{
+	// player:setMaxMana(maxMana)
+	Player* player = getPlayer(L, 1);
+	if (player) {
+		player->setMaxMana(getNumber<int32_t>(L, 2));
+		player->setMana(std::min<int32_t>(player->getMana(), player->getMaxMana()));
+		player->sendStats();
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetManaSpent(lua_State* L)
+{
+	// player:getManaSpent()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getSpentMana());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddManaSpent(lua_State* L)
+{
+	// player:addManaSpent(amount)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->addManaSpent(getNumber<uint64_t>(L, 2));
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerRemoveManaSpent(lua_State* L)
+{
+	// player:removeManaSpent(amount[, notify = true])
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->removeManaSpent(getNumber<uint64_t>(L, 2), getBoolean(L, 3, true));
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetBaseMaxHealth(lua_State* L)
+{
+	// player:getBaseMaxHealth()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getMaxHealth());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetBaseMaxMana(lua_State* L)
+{
+	// player:getBaseMaxMana()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getMaxMana());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetSkillLevel(lua_State* L)
+{
+	// player:getSkillLevel(skillType)
+	skills_t skillType = getNumber<skills_t>(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player && skillType <= SKILL_LAST) {
+		lua_pushnumber(L, player->getBaseSkill(skillType));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetEffectiveSkillLevel(lua_State* L)
+{
+	// player:getEffectiveSkillLevel(skillType)
+	skills_t skillType = getNumber<skills_t>(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player && skillType <= SKILL_LAST) {
+		lua_pushnumber(L, player->getSkillLevel(skillType));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetSkillPercent(lua_State* L)
+{
+	// player:getSkillPercent(skillType)
+	skills_t skillType = getNumber<skills_t>(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player && skillType <= SKILL_LAST) {
+		lua_pushnumber(L, player->getSkillPercent(skillType));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetSkillTries(lua_State* L)
+{
+	// player:getSkillTries(skillType)
+	skills_t skillType = getNumber<skills_t>(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player && skillType <= SKILL_LAST) {
+		lua_pushnumber(L, player->getSkillTries(skillType));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddSkillTries(lua_State* L)
+{
+	// player:addSkillTries(skillType, tries)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		skills_t skillType = getNumber<skills_t>(L, 2);
+		uint64_t tries = getNumber<uint64_t>(L, 3);
+		player->addSkillAdvance(skillType, tries);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerRemoveSkillTries(lua_State* L)
+{
+	// player:removeSkillTries(skillType, tries[, notify = true])
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		skills_t skillType = getNumber<skills_t>(L, 2);
+		uint64_t tries = getNumber<uint64_t>(L, 3);
+		player->removeSkillTries(skillType, tries, getBoolean(L, 4, true));
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetSpecialSkill(lua_State* L)
+{
+	// player:getSpecialSkill(specialSkillType)
+	SpecialSkills_t specialSkillType = getNumber<SpecialSkills_t>(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player && specialSkillType <= SPECIALSKILL_LAST) {
+		lua_pushnumber(L, player->getSpecialSkill(specialSkillType));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddSpecialSkill(lua_State* L)
+{
+	// player:addSpecialSkill(specialSkillType, value)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	SpecialSkills_t specialSkillType = getNumber<SpecialSkills_t>(L, 2);
+	if (specialSkillType > SPECIALSKILL_LAST) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	player->setVarSpecialSkill(specialSkillType, getNumber<int32_t>(L, 3));
+	player->sendSkills();
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerAddOfflineTrainingTime(lua_State* L)
+{
+	// player:addOfflineTrainingTime(time)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		int32_t time = getNumber<int32_t>(L, 2);
+		player->addOfflineTrainingTime(time);
+		player->sendStats();
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetOfflineTrainingTime(lua_State* L)
+{
+	// player:getOfflineTrainingTime()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getOfflineTrainingTime());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerRemoveOfflineTrainingTime(lua_State* L)
+{
+	// player:removeOfflineTrainingTime(time)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		int32_t time = getNumber<int32_t>(L, 2);
+		player->removeOfflineTrainingTime(time);
+		player->sendStats();
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddOfflineTrainingTries(lua_State* L)
+{
+	// player:addOfflineTrainingTries(skillType, tries)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		skills_t skillType = getNumber<skills_t>(L, 2);
+		uint64_t tries = getNumber<uint64_t>(L, 3);
+		pushBoolean(L, player->addOfflineTrainingTries(skillType, tries));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetOfflineTrainingSkill(lua_State* L)
+{
+	// player:getOfflineTrainingSkill()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getOfflineTrainingSkill());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetOfflineTrainingSkill(lua_State* L)
+{
+	// player:setOfflineTrainingSkill(skillId)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		int32_t skillId = getNumber<int32_t>(L, 2);
+		player->setOfflineTrainingSkill(skillId);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetItemCount(lua_State* L)
+{
+	// player:getItemCount(itemId[, subType = -1])
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint16_t itemId;
+	if (isNumber(L, 2)) {
+		itemId = getNumber<uint16_t>(L, 2);
+	} else {
+		itemId = Item::items.getItemIdByName(getString(L, 2));
+		if (itemId == 0) {
+			lua_pushnil(L);
+			return 1;
+		}
+	}
+
+	int32_t subType = getNumber<int32_t>(L, 3, -1);
+	lua_pushnumber(L, player->getItemTypeCount(itemId, subType));
+	return 1;
+}
+
+int luaPlayerGetItemById(lua_State* L)
+{
+	// player:getItemById(itemId, deepSearch[, subType = -1])
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint16_t itemId;
+	if (isNumber(L, 2)) {
+		itemId = getNumber<uint16_t>(L, 2);
+	} else {
+		itemId = Item::items.getItemIdByName(getString(L, 2));
+		if (itemId == 0) {
+			lua_pushnil(L);
+			return 1;
+		}
+	}
+	bool deepSearch = getBoolean(L, 3);
+	int32_t subType = getNumber<int32_t>(L, 4, -1);
+
+	Item* item = g_game->findItemOfType(player, itemId, deepSearch, subType);
+	if (item) {
+		pushUserdata<Item>(L, item);
+		setItemMetatable(L, -1, item);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetVocation(lua_State* L)
+{
+	// player:getVocation()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		pushUserdata<Vocation>(L, player->getVocation());
+		setMetatable(L, -1, "Vocation");
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetVocation(lua_State* L)
+{
+	// player:setVocation(id or name or userdata)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	Vocation* vocation;
+	if (isNumber(L, 2)) {
+		vocation = g_vocations.getVocation(getNumber<uint16_t>(L, 2));
+	} else if (lua_isstring(L, 2)) {
+		vocation = g_vocations.getVocation(g_vocations.getVocationId(getString(L, 2)));
+	} else if (lua_isuserdata(L, 2)) {
+		vocation = getUserdata<Vocation>(L, 2);
+	} else {
+		vocation = nullptr;
+	}
+
+	if (!vocation) {
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	player->setVocation(vocation->getId());
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerGetSex(lua_State* L)
+{
+	// player:getSex()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getSex());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetSex(lua_State* L)
+{
+	// player:setSex(newSex)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		PlayerSex_t newSex = getNumber<PlayerSex_t>(L, 2);
+		player->setSex(newSex);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetTown(lua_State* L)
+{
+	// player:getTown()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		pushUserdata<Town>(L, player->getTown());
+		setMetatable(L, -1, "Town");
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetTown(lua_State* L)
+{
+	// player:setTown(town)
+	Town* town = getUserdata<Town>(L, 2);
+	if (!town) {
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->setTown(town);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetGuild(lua_State* L)
+{
+	// player:getGuild()
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	Guild* guild = player->getGuild();
+	if (!guild) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	pushUserdata<Guild>(L, guild);
+	setMetatable(L, -1, "Guild");
+	return 1;
+}
+
+int luaPlayerSetGuild(lua_State* L)
+{
+	// player:setGuild(guild)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	player->setGuild(getUserdata<Guild>(L, 2));
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerGetGuildLevel(lua_State* L)
+{
+	// player:getGuildLevel()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player && player->getGuild()) {
+		lua_pushnumber(L, player->getGuildRank()->level);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetGuildLevel(lua_State* L)
+{
+	// player:setGuildLevel(level)
+	uint8_t level = getNumber<uint8_t>(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player || !player->getGuild()) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	GuildRank_ptr rank = player->getGuild()->getRankByLevel(level);
+	if (!rank) {
+		pushBoolean(L, false);
+	} else {
+		player->setGuildRank(rank);
+		pushBoolean(L, true);
+	}
+
+	return 1;
+}
+
+int luaPlayerGetGuildNick(lua_State* L)
+{
+	// player:getGuildNick()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		pushString(L, player->getGuildNick());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetGuildNick(lua_State* L)
+{
+	// player:setGuildNick(nick)
+	const std::string& nick = getString(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->setGuildNick(nick);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetGroup(lua_State* L)
+{
+	// player:getGroup()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		pushUserdata<Group>(L, player->getGroup());
+		setMetatable(L, -1, "Group");
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetGroup(lua_State* L)
+{
+	// player:setGroup(group)
+	Group* group = getUserdata<Group>(L, 2);
+	if (!group) {
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->setGroup(group);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetStamina(lua_State* L)
+{
+	// player:getStamina()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getStaminaMinutes());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetStamina(lua_State* L)
+{
+	// player:setStamina(stamina)
+	uint16_t stamina = getNumber<uint16_t>(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->setStaminaMinutes(std::min<uint16_t>(2520, stamina));
+		player->sendStats();
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetSoul(lua_State* L)
+{
+	// player:getSoul()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getSoul());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddSoul(lua_State* L)
+{
+	// player:addSoul(soulChange)
+	int32_t soulChange = getNumber<int32_t>(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->changeSoul(soulChange);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetMaxSoul(lua_State* L)
+{
+	// player:getMaxSoul()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player && player->getVocation()) {
+		lua_pushnumber(L, player->getVocation()->getSoulMax());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetBankBalance(lua_State* L)
+{
+	// player:getBankBalance()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getBankBalance());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetBankBalance(lua_State* L)
+{
+	// player:setBankBalance(bankBalance)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	int64_t balance = getNumber<int64_t>(L, 2);
+	if (balance < 0) {
+		reportErrorFunc(L, "Invalid bank balance value.");
+		lua_pushnil(L);
+		return 1;
+	}
+
+	player->setBankBalance(balance);
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerGetStorageValue(lua_State* L)
+{
+	// player:getStorageValue(key)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint32_t key = getNumber<uint32_t>(L, 2);
+	int32_t value;
+	if (player->getStorageValue(key, value)) {
+		lua_pushnumber(L, value);
+	} else {
+		lua_pushnumber(L, -1);
+	}
+	return 1;
+}
+
+int luaPlayerSetStorageValue(lua_State* L)
+{
+	// player:setStorageValue(key, value)
+	int32_t value = getNumber<int32_t>(L, 3);
+	uint32_t key = getNumber<uint32_t>(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (IS_IN_KEYRANGE(key, RESERVED_RANGE)) {
+		reportErrorFunc(L, fmt::format("Accessing reserved range: {:d}", key));
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	if (player) {
+		player->addStorageValue(key, value);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddItem(lua_State* L)
+{
+	// player:addItem(itemId[, count = 1[, canDropOnMap = true[, subType = 1[, slot = CONST_SLOT_WHEREEVER]]]])
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	uint16_t itemId;
+	if (isNumber(L, 2)) {
+		itemId = getNumber<uint16_t>(L, 2);
+	} else {
+		itemId = Item::items.getItemIdByName(getString(L, 2));
+		if (itemId == 0) {
+			lua_pushnil(L);
+			return 1;
+		}
+	}
+
+	int32_t count = getNumber<int32_t>(L, 3, 1);
+	int32_t subType = getNumber<int32_t>(L, 5, 1);
+
+	const ItemType& it = Item::items[itemId];
+
+	int32_t itemCount = 1;
+	int parameters = lua_gettop(L);
+	if (parameters >= 5) {
+		itemCount = std::max<int32_t>(1, count);
+	} else if (it.hasSubType()) {
+		if (it.stackable) {
+			itemCount = std::ceil(count / 100.f);
+		}
+
+		subType = count;
+	} else {
+		itemCount = std::max<int32_t>(1, count);
+	}
+
+	bool hasTable = itemCount > 1;
+	if (hasTable) {
+		lua_newtable(L);
+	} else if (itemCount == 0) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	bool canDropOnMap = getBoolean(L, 4, true);
+	slots_t slot = getNumber<slots_t>(L, 6, CONST_SLOT_WHEREEVER);
+	for (int32_t i = 1; i <= itemCount; ++i) {
+		int32_t stackCount = subType;
+		if (it.stackable) {
+			stackCount = std::min<int32_t>(stackCount, 100);
+			subType -= stackCount;
+		}
+
+		Item* item = Item::CreateItem(itemId, stackCount);
+		if (!item) {
+			if (!hasTable) {
+				lua_pushnil(L);
+			}
+			return 1;
+		}
+
+		ReturnValue ret = g_game->internalPlayerAddItem(player, item, canDropOnMap, slot);
+		if (ret != RETURNVALUE_NOERROR) {
+			delete item;
+			if (!hasTable) {
+				lua_pushnil(L);
+			}
+			return 1;
+		}
+
+		if (hasTable) {
+			lua_pushnumber(L, i);
+			pushUserdata<Item>(L, item);
+			setItemMetatable(L, -1, item);
+			lua_settable(L, -3);
+		} else {
+			pushUserdata<Item>(L, item);
+			setItemMetatable(L, -1, item);
+		}
+	}
+	return 1;
+}
+
+int luaPlayerAddItemEx(lua_State* L)
+{
+	// player:addItemEx(item[, canDropOnMap = false[, index = INDEX_WHEREEVER[, flags = 0]]]) player:addItemEx(item[,
+	// canDropOnMap = true[, slot = CONST_SLOT_WHEREEVER]])
+	Item* item = getUserdata<Item>(L, 2);
+	if (!item) {
+		reportErrorFunc(L, getErrorDesc(LUA_ERROR_ITEM_NOT_FOUND));
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	if (item->getParent() != VirtualCylinder::virtualCylinder) {
+		reportErrorFunc(L, "Item already has a parent");
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	bool canDropOnMap = getBoolean(L, 3, false);
+	ReturnValue returnValue;
+	if (canDropOnMap) {
+		slots_t slot = getNumber<slots_t>(L, 4, CONST_SLOT_WHEREEVER);
+		returnValue = g_game->internalPlayerAddItem(player, item, true, slot);
+	} else {
+		int32_t index = getNumber<int32_t>(L, 4, INDEX_WHEREEVER);
+		uint32_t flags = getNumber<uint32_t>(L, 5, 0);
+		returnValue = g_game->internalAddItem(player, item, index, flags);
+	}
+
+	if (returnValue == RETURNVALUE_NOERROR) {
+		ScriptEnvironment::removeTempItem(item);
+	}
+	lua_pushnumber(L, returnValue);
+	return 1;
+}
+
+int luaPlayerRemoveItem(lua_State* L)
+{
+	// player:removeItem(itemId, count[, subType = -1[, ignoreEquipped = false]])
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint16_t itemId;
+	if (isNumber(L, 2)) {
+		itemId = getNumber<uint16_t>(L, 2);
+	} else {
+		itemId = Item::items.getItemIdByName(getString(L, 2));
+		if (itemId == 0) {
+			lua_pushnil(L);
+			return 1;
+		}
+	}
+
+	uint32_t count = getNumber<uint32_t>(L, 3);
+	int32_t subType = getNumber<int32_t>(L, 4, -1);
+	bool ignoreEquipped = getBoolean(L, 5, false);
+	pushBoolean(L, player->removeItemOfType(itemId, count, subType, ignoreEquipped));
+	return 1;
+}
+
+int luaPlayerSendSupplyUsed(lua_State* L)
+{
+	// player:sendSupplyUsed(item)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		reportErrorFunc(L, getErrorDesc(LUA_ERROR_PLAYER_NOT_FOUND));
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	Item* item = getUserdata<Item>(L, 2);
+	if (!item) {
+		reportErrorFunc(L, getErrorDesc(LUA_ERROR_ITEM_NOT_FOUND));
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	player->sendSupplyUsed(item->getClientID());
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerGetMoney(lua_State* L)
+{
+	// player:getMoney()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getMoney());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddMoney(lua_State* L)
+{
+	// player:addMoney(money)
+	uint64_t money = getNumber<uint64_t>(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		g_game->addMoney(player, money);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerRemoveMoney(lua_State* L)
+{
+	// player:removeMoney(money)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		uint64_t money = getNumber<uint64_t>(L, 2);
+		pushBoolean(L, g_game->removeMoney(player, money));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerShowTextDialog(lua_State* L)
+{
+	// player:showTextDialog(id or name or userdata[, text[, canWrite[, length]]])
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	int32_t length = getNumber<int32_t>(L, 5, -1);
+	bool canWrite = getBoolean(L, 4, false);
+	std::string text;
+
+	int parameters = lua_gettop(L);
+	if (parameters >= 3) {
+		text = getString(L, 3);
+	}
+
+	Item* item;
+	if (isNumber(L, 2)) {
+		item = Item::CreateItem(getNumber<uint16_t>(L, 2));
+	} else if (lua_isstring(L, 2)) {
+		item = Item::CreateItem(Item::items.getItemIdByName(getString(L, 2)));
+	} else if (lua_isuserdata(L, 2)) {
+		if (getUserdataType(L, 2) != LuaData_Item) {
+			pushBoolean(L, false);
+			return 1;
+		}
+
+		item = getUserdata<Item>(L, 2);
+	} else {
+		item = nullptr;
+	}
+
+	if (!item) {
+		reportErrorFunc(L, getErrorDesc(LUA_ERROR_ITEM_NOT_FOUND));
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	if (length < 0) {
+		length = Item::items[item->getID()].maxTextLen;
+	}
+
+	if (!text.empty()) {
+		item->setText(text);
+		length = std::max<int32_t>(text.size(), length);
+	}
+
+	item->setParent(player);
+	player->incrementWindowTextId();
+	player->setWriteItem(item, length);
+	player->sendTextWindow(item, length, canWrite);
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerSendTextMessage(lua_State* L)
+{
+	// player:sendTextMessage(type, text[, position, primaryValue = 0, primaryColor = TEXTCOLOR_NONE[, secondaryValue =
+	// 0, secondaryColor = TEXTCOLOR_NONE]]) player:sendTextMessage(type, text, channelId)
+
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	int parameters = lua_gettop(L);
+
+	TextMessage message(getNumber<MessageClasses>(L, 2), getString(L, 3));
+	if (parameters == 4) {
+		uint16_t channelId = getNumber<uint16_t>(L, 4);
+		ChatChannel* channel = g_chat->getChannel(*player, channelId);
+		if (!channel || !channel->hasUser(*player)) {
+			pushBoolean(L, false);
+			return 1;
+		}
+		message.channelId = channelId;
+	} else {
+		if (parameters >= 6) {
+			message.position = getPosition(L, 4);
+			message.primary.value = getNumber<int32_t>(L, 5);
+			message.primary.color = getNumber<TextColor_t>(L, 6);
+		}
+
+		if (parameters >= 8) {
+			message.secondary.value = getNumber<int32_t>(L, 7);
+			message.secondary.color = getNumber<TextColor_t>(L, 8);
+		}
+	}
+
+	player->sendTextMessage(message);
+	pushBoolean(L, true);
+
+	return 1;
+}
+
+int luaPlayerSendChannelMessage(lua_State* L)
+{
+	// player:sendChannelMessage(author, text, type, channelId)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint16_t channelId = getNumber<uint16_t>(L, 5);
+	SpeakClasses type = getNumber<SpeakClasses>(L, 4);
+	const std::string& text = getString(L, 3);
+	const std::string& author = getString(L, 2);
+	player->sendChannelMessage(author, text, type, channelId);
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerSendPrivateMessage(lua_State* L)
+{
+	// player:sendPrivateMessage(speaker, text[, type])
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	const Player* speaker = getUserdata<const Player>(L, 2);
+	const std::string& text = getString(L, 3);
+	SpeakClasses type = getNumber<SpeakClasses>(L, 4, TALKTYPE_PRIVATE_FROM);
+	player->sendPrivateMessage(speaker, type, text);
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerChannelSay(lua_State* L)
+{
+	// player:channelSay(speaker, type, text, channelId)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	Creature* speaker = getCreature(L, 2);
+	SpeakClasses type = getNumber<SpeakClasses>(L, 3);
+	const std::string& text = getString(L, 4);
+	uint16_t channelId = getNumber<uint16_t>(L, 5);
+	player->sendToChannel(speaker, type, text, channelId);
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerOpenChannel(lua_State* L)
+{
+	// player:openChannel(channelId)
+	uint16_t channelId = getNumber<uint16_t>(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		g_game->playerOpenChannel(player->getID(), channelId);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetSlotItem(lua_State* L)
+{
+	// player:getSlotItem(slot)
+	const Player* player = getUserdata<const Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint32_t slot = getNumber<uint32_t>(L, 2);
+	Thing* thing = player->getThing(slot);
+	if (!thing) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	Item* item = thing->getItem();
+	if (item) {
+		pushUserdata<Item>(L, item);
+		setItemMetatable(L, -1, item);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetParty(lua_State* L)
+{
+	// player:getParty()
+	const Player* player = getUserdata<const Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	Party* party = player->getParty();
+	if (party) {
+		pushUserdata<Party>(L, party);
+		setMetatable(L, -1, "Party");
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddOutfit(lua_State* L)
+{
+	// player:addOutfit(lookType)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->addOutfit(getNumber<uint16_t>(L, 2), 0);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddOutfitAddon(lua_State* L)
+{
+	// player:addOutfitAddon(lookType, addon)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		uint16_t lookType = getNumber<uint16_t>(L, 2);
+		uint8_t addon = getNumber<uint8_t>(L, 3);
+		player->addOutfit(lookType, addon);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerRemoveOutfit(lua_State* L)
+{
+	// player:removeOutfit(lookType)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		uint16_t lookType = getNumber<uint16_t>(L, 2);
+		pushBoolean(L, player->removeOutfit(lookType));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerRemoveOutfitAddon(lua_State* L)
+{
+	// player:removeOutfitAddon(lookType, addon)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		uint16_t lookType = getNumber<uint16_t>(L, 2);
+		uint8_t addon = getNumber<uint8_t>(L, 3);
+		pushBoolean(L, player->removeOutfitAddon(lookType, addon));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerHasOutfit(lua_State* L)
+{
+	// player:hasOutfit(lookType[, addon = 0])
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		uint16_t lookType = getNumber<uint16_t>(L, 2);
+		uint8_t addon = getNumber<uint8_t>(L, 3, 0);
+		pushBoolean(L, player->hasOutfit(lookType, addon));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerCanWearOutfit(lua_State* L)
+{
+	// player:canWearOutfit(lookType[, addon = 0])
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		uint16_t lookType = getNumber<uint16_t>(L, 2);
+		uint8_t addon = getNumber<uint8_t>(L, 3, 0);
+		pushBoolean(L, player->canWear(lookType, addon));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSendOutfitWindow(lua_State* L)
+{
+	// player:sendOutfitWindow()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->sendOutfitWindow();
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSendEditPodium(lua_State* L)
+{
+	// player:sendEditPodium(item)
+	Player* player = getUserdata<Player>(L, 1);
+	Item* item = getUserdata<Item>(L, 2);
+	if (player && item) {
+		player->sendPodiumWindow(item);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddMount(lua_State* L)
+{
+	// player:addMount(mountId or mountName)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint8_t mountId;
+	if (isNumber(L, 2)) {
+		mountId = getNumber<uint8_t>(L, 2);
+	} else {
+		Mount* mount = g_game->mounts.getMountByName(getString(L, 2));
+		if (!mount) {
+			lua_pushnil(L);
+			return 1;
+		}
+		mountId = mount->id;
+	}
+	pushBoolean(L, player->tameMount(mountId));
+	return 1;
+}
+
+int luaPlayerRemoveMount(lua_State* L)
+{
+	// player:removeMount(mountId or mountName)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint8_t mountId;
+	if (isNumber(L, 2)) {
+		mountId = getNumber<uint8_t>(L, 2);
+	} else {
+		Mount* mount = g_game->mounts.getMountByName(getString(L, 2));
+		if (!mount) {
+			lua_pushnil(L);
+			return 1;
+		}
+		mountId = mount->id;
+	}
+	pushBoolean(L, player->untameMount(mountId));
+	return 1;
+}
+
+int luaPlayerHasMount(lua_State* L)
+{
+	// player:hasMount(mountId or mountName)
+	const Player* player = getUserdata<const Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	Mount* mount = nullptr;
+	if (isNumber(L, 2)) {
+		mount = g_game->mounts.getMountByID(getNumber<uint8_t>(L, 2));
+	} else {
+		mount = g_game->mounts.getMountByName(getString(L, 2));
+	}
+
+	if (mount) {
+		pushBoolean(L, player->hasMount(mount));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetPremiumEndsAt(lua_State* L)
+{
+	// player:getPremiumEndsAt()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getPremiumTime());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSetPremiumEndsAt(lua_State* L)
+{
+	// player:setPremiumEndsAt(timestamp)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	time_t timestamp = getNumber<time_t>(L, 2);
+
+	player->setPremiumTime(timestamp);
+	IOLoginData::updatePremiumTime(player->getAccount(), timestamp);
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerHasBlessing(lua_State* L)
+{
+	// player:hasBlessing(blessing)
+	uint8_t blessing = getNumber<uint8_t>(L, 2) - 1;
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		pushBoolean(L, player->hasBlessing(blessing));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddBlessing(lua_State* L)
+{
+	// player:addBlessing(blessing)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint8_t blessing = getNumber<uint8_t>(L, 2) - 1;
+	if (player->hasBlessing(blessing)) {
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	player->addBlessing(blessing);
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerRemoveBlessing(lua_State* L)
+{
+	// player:removeBlessing(blessing)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint8_t blessing = getNumber<uint8_t>(L, 2) - 1;
+	if (!player->hasBlessing(blessing)) {
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	player->removeBlessing(blessing);
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerCanLearnSpell(lua_State* L)
+{
+	// player:canLearnSpell(spellName)
+	const Player* player = getUserdata<const Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	const std::string& spellName = getString(L, 2);
+	InstantSpell* spell = g_spells->getInstantSpellByName(spellName);
+	if (!spell) {
+		reportErrorFunc(L, "Spell \"" + spellName + "\" not found");
+		pushBoolean(L, false);
+		return 1;
+	}
+
+	if (player->hasFlag(PlayerFlag_IgnoreSpellCheck)) {
+		pushBoolean(L, true);
+		return 1;
+	}
+
+	const auto& vocMap = spell->getVocMap();
+	if (vocMap.count(player->getVocationId()) == 0) {
+		pushBoolean(L, false);
+	} else if (player->getLevel() < spell->getLevel()) {
+		pushBoolean(L, false);
+	} else if (player->getMagicLevel() < spell->getMagicLevel()) {
+		pushBoolean(L, false);
+	} else {
+		pushBoolean(L, true);
+	}
+	return 1;
+}
+
+int luaPlayerLearnSpell(lua_State* L)
+{
+	// player:learnSpell(spellName)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		const std::string& spellName = getString(L, 2);
+		player->learnInstantSpell(spellName);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerForgetSpell(lua_State* L)
+{
+	// player:forgetSpell(spellName)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		const std::string& spellName = getString(L, 2);
+		player->forgetInstantSpell(spellName);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerHasLearnedSpell(lua_State* L)
+{
+	// player:hasLearnedSpell(spellName)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		const std::string& spellName = getString(L, 2);
+		pushBoolean(L, player->hasLearnedInstantSpell(spellName));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSendTutorial(lua_State* L)
+{
+	// player:sendTutorial(tutorialId)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		uint8_t tutorialId = getNumber<uint8_t>(L, 2);
+		player->sendTutorial(tutorialId);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerAddMapMark(lua_State* L)
+{
+	// player:addMapMark(position, type, description)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		const Position& position = getPosition(L, 2);
+		uint8_t type = getNumber<uint8_t>(L, 3);
+		const std::string& description = getString(L, 4);
+		player->sendAddMarker(position, type, description);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSave(lua_State* L)
+{
+	// player:save()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->setLoginPosition(player->getPosition());
+		pushBoolean(L, IOLoginData::savePlayer(player));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerPopupFYI(lua_State* L)
+{
+	// player:popupFYI(message)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		const std::string& message = getString(L, 2);
+		player->sendFYIBox(message);
+		pushBoolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerIsPzLocked(lua_State* L)
+{
+	// player:isPzLocked()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		pushBoolean(L, player->isPzLocked());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetClient(lua_State* L)
+{
+	// player:getClient()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_createtable(L, 0, 2);
+		setField(L, "version", player->getProtocolVersion());
+		setField(L, "os", player->getOperatingSystem());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetHouse(lua_State* L)
+{
+	// player:getHouse()
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	House* house = g_game->map.houses.getHouseByPlayerId(player->getGUID());
+	if (house) {
+		pushUserdata<House>(L, house);
+		setMetatable(L, -1, "House");
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerSendHouseWindow(lua_State* L)
+{
+	// player:sendHouseWindow(house, listId)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	House* house = getUserdata<House>(L, 2);
+	if (!house) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint32_t listId = getNumber<uint32_t>(L, 3);
+	player->sendHouseWindow(house, listId);
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerSetEditHouse(lua_State* L)
+{
+	// player:setEditHouse(house, listId)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	House* house = getUserdata<House>(L, 2);
+	if (!house) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	uint32_t listId = getNumber<uint32_t>(L, 3);
+	player->setEditHouse(house, listId);
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerSetGhostMode(lua_State* L)
+{
+	// player:setGhostMode(enabled[, magicEffect = CONST_ME_TELEPORT])
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	bool enabled = getBoolean(L, 2);
+	if (player->isInGhostMode() == enabled) {
+		pushBoolean(L, true);
+		return 1;
+	}
+
+	MagicEffectClasses magicEffect = getNumber<MagicEffectClasses>(L, 3, CONST_ME_TELEPORT);
+
+	player->switchGhostMode();
+
+	Tile* tile = player->getTile();
+	const Position& position = player->getPosition();
+	const bool isInvisible = player->isInvisible();
+
+	SpectatorVec spectators;
+	g_game->map.getSpectators(spectators, position, true, true);
+	for (Creature* spectator : spectators) {
+		Player* tmpPlayer = spectator->getPlayer();
+		if (tmpPlayer != player && !tmpPlayer->isAccessPlayer()) {
+			if (enabled) {
+				tmpPlayer->sendRemoveTileCreature(player, position, tile->getClientIndexOfCreature(tmpPlayer, player));
+			} else {
+				tmpPlayer->sendCreatureAppear(player, position, magicEffect);
+			}
+		} else {
+			if (isInvisible) {
+				continue;
+			}
+
+			tmpPlayer->sendCreatureChangeVisible(player, !enabled);
+		}
+	}
+
+	if (player->isInGhostMode()) {
+		for (const auto& it : g_game->getPlayers()) {
+			if (!it.second->isAccessPlayer()) {
+				it.second->notifyStatusChange(player, VIPSTATUS_OFFLINE);
+			}
+		}
+		IOLoginData::updateOnlineStatus(player->getGUID(), false);
+	} else {
+		for (const auto& it : g_game->getPlayers()) {
+			if (!it.second->isAccessPlayer()) {
+				it.second->notifyStatusChange(player, VIPSTATUS_ONLINE);
+			}
+		}
+		IOLoginData::updateOnlineStatus(player->getGUID(), true);
+	}
+	pushBoolean(L, true);
+	return 1;
+}
+
+int luaPlayerGetContainerId(lua_State* L)
+{
+	// player:getContainerId(container)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	Container* container = getUserdata<Container>(L, 2);
+	if (container) {
+		lua_pushnumber(L, player->getContainerID(container));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetContainerById(lua_State* L)
+{
+	// player:getContainerById(id)
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	Container* container = player->getContainerByID(getNumber<uint8_t>(L, 2));
+	if (container) {
+		pushUserdata<Container>(L, container);
+		setMetatable(L, -1, "Container");
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetContainerIndex(lua_State* L)
+{
+	// player:getContainerIndex(id)
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getContainerIndex(getNumber<uint8_t>(L, 2)));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetInstantSpells(lua_State* L)
+{
+	// player:getInstantSpells()
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	std::vector<const InstantSpell*> spells;
+	for (auto& spell : g_spells->getInstantSpells()) {
+		if (spell.second.canCast(player)) {
+			spells.push_back(&spell.second);
+		}
+	}
+
+	lua_createtable(L, spells.size(), 0);
+
+	int index = 0;
+	for (auto spell : spells) {
+		pushInstantSpell(L, *spell);
+		lua_rawseti(L, -2, ++index);
+	}
+	return 1;
+}
+
+int luaPlayerCanCast(lua_State* L)
+{
+	// player:canCast(spell)
+	Player* player = getUserdata<Player>(L, 1);
+	InstantSpell* spell = getUserdata<InstantSpell>(L, 2);
+	if (player && spell) {
+		pushBoolean(L, spell->canCast(player));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerHasChaseMode(lua_State* L)
+{
+	// player:hasChaseMode()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		pushBoolean(L, player->getChaseMode());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerHasSecureMode(lua_State* L)
+{
+	// player:hasSecureMode()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		pushBoolean(L, player->getSecureMode());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetFightMode(lua_State* L)
+{
+	// player:getFightMode()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushnumber(L, player->getFightMode());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int luaPlayerGetStoreInbox(lua_State* L)
+{
+	// player:getStoreInbox()
+	Player* player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	Container* storeInbox = player->getStoreInbox();
+	if (!storeInbox) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	pushUserdata<Container>(L, storeInbox);
+	setMetatable(L, -1, "Container");
+	return 1;
+}
+
+int luaPlayerIsNearDepotBox(lua_State* L)
+{
+	// player:isNearDepotBox()
+	const Player* const player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	pushBoolean(L, player->isNearDepotBox());
+	return 1;
+}
+
+int luaPlayerGetIdleTime(lua_State* L)
+{
+	// player:getIdleTime()
+	const Player* const player = getUserdata<Player>(L, 1);
+	if (!player) {
+		lua_pushnil(L);
+		return 1;
+	}
+
+	lua_pushnumber(L, player->getIdleTime());
+	return 1;
+}
+
+void registerFunctions(LuaScriptInterface& lsi)
+{
+	lsi.registerClass("Player", "Creature", luaPlayerCreate);
+	lsi.registerMetaMethod("Player", "__eq", luaUserdataCompare);
+
+	lsi.registerMethod("Player", "isPlayer", luaPlayerIsPlayer);
+
+	lsi.registerMethod("Player", "getGuid", luaPlayerGetGuid);
+	lsi.registerMethod("Player", "getIp", luaPlayerGetIp);
+	lsi.registerMethod("Player", "getAccountId", luaPlayerGetAccountId);
+	lsi.registerMethod("Player", "getLastLoginSaved", luaPlayerGetLastLoginSaved);
+	lsi.registerMethod("Player", "getLastLogout", luaPlayerGetLastLogout);
+
+	lsi.registerMethod("Player", "getAccountType", luaPlayerGetAccountType);
+	lsi.registerMethod("Player", "setAccountType", luaPlayerSetAccountType);
+
+	lsi.registerMethod("Player", "getCapacity", luaPlayerGetCapacity);
+	lsi.registerMethod("Player", "setCapacity", luaPlayerSetCapacity);
+
+	lsi.registerMethod("Player", "getFreeCapacity", luaPlayerGetFreeCapacity);
+
+	lsi.registerMethod("Player", "getDepotChest", luaPlayerGetDepotChest);
+	lsi.registerMethod("Player", "getInbox", luaPlayerGetInbox);
+
+	lsi.registerMethod("Player", "getSkullTime", luaPlayerGetSkullTime);
+	lsi.registerMethod("Player", "setSkullTime", luaPlayerSetSkullTime);
+	lsi.registerMethod("Player", "getDeathPenalty", luaPlayerGetDeathPenalty);
+
+	lsi.registerMethod("Player", "getExperience", luaPlayerGetExperience);
+	lsi.registerMethod("Player", "addExperience", luaPlayerAddExperience);
+	lsi.registerMethod("Player", "removeExperience", luaPlayerRemoveExperience);
+	lsi.registerMethod("Player", "getLevel", luaPlayerGetLevel);
+
+	lsi.registerMethod("Player", "getMagicLevel", luaPlayerGetMagicLevel);
+	lsi.registerMethod("Player", "getBaseMagicLevel", luaPlayerGetBaseMagicLevel);
+	lsi.registerMethod("Player", "getMana", luaPlayerGetMana);
+	lsi.registerMethod("Player", "addMana", luaPlayerAddMana);
+	lsi.registerMethod("Player", "getMaxMana", luaPlayerGetMaxMana);
+	lsi.registerMethod("Player", "setMaxMana", luaPlayerSetMaxMana);
+	lsi.registerMethod("Player", "getManaSpent", luaPlayerGetManaSpent);
+	lsi.registerMethod("Player", "addManaSpent", luaPlayerAddManaSpent);
+	lsi.registerMethod("Player", "removeManaSpent", luaPlayerRemoveManaSpent);
+
+	lsi.registerMethod("Player", "getBaseMaxHealth", luaPlayerGetBaseMaxHealth);
+	lsi.registerMethod("Player", "getBaseMaxMana", luaPlayerGetBaseMaxMana);
+
+	lsi.registerMethod("Player", "getSkillLevel", luaPlayerGetSkillLevel);
+	lsi.registerMethod("Player", "getEffectiveSkillLevel", luaPlayerGetEffectiveSkillLevel);
+	lsi.registerMethod("Player", "getSkillPercent", luaPlayerGetSkillPercent);
+	lsi.registerMethod("Player", "getSkillTries", luaPlayerGetSkillTries);
+	lsi.registerMethod("Player", "addSkillTries", luaPlayerAddSkillTries);
+	lsi.registerMethod("Player", "removeSkillTries", luaPlayerRemoveSkillTries);
+	lsi.registerMethod("Player", "getSpecialSkill", luaPlayerGetSpecialSkill);
+	lsi.registerMethod("Player", "addSpecialSkill", luaPlayerAddSpecialSkill);
+
+	lsi.registerMethod("Player", "addOfflineTrainingTime", luaPlayerAddOfflineTrainingTime);
+	lsi.registerMethod("Player", "getOfflineTrainingTime", luaPlayerGetOfflineTrainingTime);
+	lsi.registerMethod("Player", "removeOfflineTrainingTime", luaPlayerRemoveOfflineTrainingTime);
+
+	lsi.registerMethod("Player", "addOfflineTrainingTries", luaPlayerAddOfflineTrainingTries);
+
+	lsi.registerMethod("Player", "getOfflineTrainingSkill", luaPlayerGetOfflineTrainingSkill);
+	lsi.registerMethod("Player", "setOfflineTrainingSkill", luaPlayerSetOfflineTrainingSkill);
+
+	lsi.registerMethod("Player", "getItemCount", luaPlayerGetItemCount);
+	lsi.registerMethod("Player", "getItemById", luaPlayerGetItemById);
+
+	lsi.registerMethod("Player", "getVocation", luaPlayerGetVocation);
+	lsi.registerMethod("Player", "setVocation", luaPlayerSetVocation);
+
+	lsi.registerMethod("Player", "getSex", luaPlayerGetSex);
+	lsi.registerMethod("Player", "setSex", luaPlayerSetSex);
+
+	lsi.registerMethod("Player", "getTown", luaPlayerGetTown);
+	lsi.registerMethod("Player", "setTown", luaPlayerSetTown);
+
+	lsi.registerMethod("Player", "getGuild", luaPlayerGetGuild);
+	lsi.registerMethod("Player", "setGuild", luaPlayerSetGuild);
+
+	lsi.registerMethod("Player", "getGuildLevel", luaPlayerGetGuildLevel);
+	lsi.registerMethod("Player", "setGuildLevel", luaPlayerSetGuildLevel);
+
+	lsi.registerMethod("Player", "getGuildNick", luaPlayerGetGuildNick);
+	lsi.registerMethod("Player", "setGuildNick", luaPlayerSetGuildNick);
+
+	lsi.registerMethod("Player", "getGroup", luaPlayerGetGroup);
+	lsi.registerMethod("Player", "setGroup", luaPlayerSetGroup);
+
+	lsi.registerMethod("Player", "getStamina", luaPlayerGetStamina);
+	lsi.registerMethod("Player", "setStamina", luaPlayerSetStamina);
+
+	lsi.registerMethod("Player", "getSoul", luaPlayerGetSoul);
+	lsi.registerMethod("Player", "addSoul", luaPlayerAddSoul);
+	lsi.registerMethod("Player", "getMaxSoul", luaPlayerGetMaxSoul);
+
+	lsi.registerMethod("Player", "getBankBalance", luaPlayerGetBankBalance);
+	lsi.registerMethod("Player", "setBankBalance", luaPlayerSetBankBalance);
+
+	lsi.registerMethod("Player", "getStorageValue", luaPlayerGetStorageValue);
+	lsi.registerMethod("Player", "setStorageValue", luaPlayerSetStorageValue);
+
+	lsi.registerMethod("Player", "addItem", luaPlayerAddItem);
+	lsi.registerMethod("Player", "addItemEx", luaPlayerAddItemEx);
+	lsi.registerMethod("Player", "removeItem", luaPlayerRemoveItem);
+	lsi.registerMethod("Player", "sendSupplyUsed", luaPlayerSendSupplyUsed);
+
+	lsi.registerMethod("Player", "getMoney", luaPlayerGetMoney);
+	lsi.registerMethod("Player", "addMoney", luaPlayerAddMoney);
+	lsi.registerMethod("Player", "removeMoney", luaPlayerRemoveMoney);
+
+	lsi.registerMethod("Player", "showTextDialog", luaPlayerShowTextDialog);
+
+	lsi.registerMethod("Player", "sendTextMessage", luaPlayerSendTextMessage);
+	lsi.registerMethod("Player", "sendChannelMessage", luaPlayerSendChannelMessage);
+	lsi.registerMethod("Player", "sendPrivateMessage", luaPlayerSendPrivateMessage);
+	lsi.registerMethod("Player", "channelSay", luaPlayerChannelSay);
+	lsi.registerMethod("Player", "openChannel", luaPlayerOpenChannel);
+
+	lsi.registerMethod("Player", "getSlotItem", luaPlayerGetSlotItem);
+
+	lsi.registerMethod("Player", "getParty", luaPlayerGetParty);
+
+	lsi.registerMethod("Player", "addOutfit", luaPlayerAddOutfit);
+	lsi.registerMethod("Player", "addOutfitAddon", luaPlayerAddOutfitAddon);
+	lsi.registerMethod("Player", "removeOutfit", luaPlayerRemoveOutfit);
+	lsi.registerMethod("Player", "removeOutfitAddon", luaPlayerRemoveOutfitAddon);
+	lsi.registerMethod("Player", "hasOutfit", luaPlayerHasOutfit);
+	lsi.registerMethod("Player", "canWearOutfit", luaPlayerCanWearOutfit);
+	lsi.registerMethod("Player", "sendOutfitWindow", luaPlayerSendOutfitWindow);
+
+	lsi.registerMethod("Player", "sendEditPodium", luaPlayerSendEditPodium);
+
+	lsi.registerMethod("Player", "addMount", luaPlayerAddMount);
+	lsi.registerMethod("Player", "removeMount", luaPlayerRemoveMount);
+	lsi.registerMethod("Player", "hasMount", luaPlayerHasMount);
+
+	lsi.registerMethod("Player", "getPremiumEndsAt", luaPlayerGetPremiumEndsAt);
+	lsi.registerMethod("Player", "setPremiumEndsAt", luaPlayerSetPremiumEndsAt);
+
+	lsi.registerMethod("Player", "hasBlessing", luaPlayerHasBlessing);
+	lsi.registerMethod("Player", "addBlessing", luaPlayerAddBlessing);
+	lsi.registerMethod("Player", "removeBlessing", luaPlayerRemoveBlessing);
+
+	lsi.registerMethod("Player", "canLearnSpell", luaPlayerCanLearnSpell);
+	lsi.registerMethod("Player", "learnSpell", luaPlayerLearnSpell);
+	lsi.registerMethod("Player", "forgetSpell", luaPlayerForgetSpell);
+	lsi.registerMethod("Player", "hasLearnedSpell", luaPlayerHasLearnedSpell);
+
+	lsi.registerMethod("Player", "sendTutorial", luaPlayerSendTutorial);
+	lsi.registerMethod("Player", "addMapMark", luaPlayerAddMapMark);
+
+	lsi.registerMethod("Player", "save", luaPlayerSave);
+	lsi.registerMethod("Player", "popupFYI", luaPlayerPopupFYI);
+
+	lsi.registerMethod("Player", "isPzLocked", luaPlayerIsPzLocked);
+
+	lsi.registerMethod("Player", "getClient", luaPlayerGetClient);
+
+	lsi.registerMethod("Player", "getHouse", luaPlayerGetHouse);
+	lsi.registerMethod("Player", "sendHouseWindow", luaPlayerSendHouseWindow);
+	lsi.registerMethod("Player", "setEditHouse", luaPlayerSetEditHouse);
+
+	lsi.registerMethod("Player", "setGhostMode", luaPlayerSetGhostMode);
+
+	lsi.registerMethod("Player", "getContainerId", luaPlayerGetContainerId);
+	lsi.registerMethod("Player", "getContainerById", luaPlayerGetContainerById);
+	lsi.registerMethod("Player", "getContainerIndex", luaPlayerGetContainerIndex);
+
+	lsi.registerMethod("Player", "getInstantSpells", luaPlayerGetInstantSpells);
+	lsi.registerMethod("Player", "canCast", luaPlayerCanCast);
+
+	lsi.registerMethod("Player", "hasChaseMode", luaPlayerHasChaseMode);
+	lsi.registerMethod("Player", "hasSecureMode", luaPlayerHasSecureMode);
+	lsi.registerMethod("Player", "getFightMode", luaPlayerGetFightMode);
+
+	lsi.registerMethod("Player", "getStoreInbox", luaPlayerGetStoreInbox);
+
+	lsi.registerMethod("Player", "isNearDepotBox", luaPlayerIsNearDepotBox);
+
+	lsi.registerMethod("Player", "getIdleTime", luaPlayerGetIdleTime);
+}
+
+} // namespace
+
+registerLuaModule("player", registerFunctions);
